pyproject.toml
[tool.poetry]
name = "drf-for-student"
version = "0.1.0"
description = ""
authors = ["egor <sharapov.egor00@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
djangorestframework = "^3.15.1"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
pillow = "^10.3.0"
ipython = "^8.24.0"
pytils = "^0.4.1"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^7.3.0"}
django-filter = "^24.2"
djangorestframework-simplejwt = "^5.3.1"
coverage = "^7.5.3"
drf-yasg = "^1.21.7"
setuptools = "^70.0.0"



[tool.poetry.group.lint.dependencies]
black = "^23.10.1"
isort = "^5.12.0"
mypy = "^1.6.1"
flake8 = "^6.1.0"

[tool.black]
# максимальная длина строки
line-length = 119
# файлы, которые не нужно форматировать
exclude = '''
(
  /(
      \.eggs         # исключить несколько общих каталогов в
    | \.git          # корне проекта
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py           # также отдельно исключить файл с именем foo.py в
                     # корне проекта
)
'''

[tool.isort]
# максимальная длина строки
line_length = 119

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
